chore: rename view mode toggle component
